Move `Subscribe` to `annotation` package
