+++ server/src/main/java/org/spine3/server/event/EventBus.java
@@ -30,7 +30,7 @@ import org.slf4j.LoggerFactory;
 import org.spine3.annotation.Internal;
 import org.spine3.base.Event;
 import org.spine3.base.Response;
-import org.spine3.base.Subscribe;
+import org.spine3.annotation.Subscribe;
 import org.spine3.envelope.EventEnvelope;
 import org.spine3.io.StreamObservers;
 import org.spine3.server.event.enrich.EventEnricher;
