+++ server/src/main/java/org/spine3/server/reflect/EventSubscriberMethod.java
@@ -24,7 +24,7 @@ import com.google.common.base.Predicate;
 import com.google.common.collect.ImmutableSet;
 import com.google.protobuf.Message;
 import org.spine3.base.EventContext;
-import org.spine3.base.Subscribe;
+import org.spine3.annotation.Subscribe;
 import org.spine3.type.EventClass;
 
 import javax.annotation.CheckReturnValue;
