+++ server/src/test/java/org/spine3/server/QueryServiceShould.java
@@ -26,7 +26,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.spine3.base.EventContext;
 import org.spine3.base.Responses;
-import org.spine3.base.Subscribe;
+import org.spine3.annotation.Subscribe;
 import org.spine3.client.Query;
 import org.spine3.client.QueryResponse;
 import org.spine3.server.projection.Projection;
