+++ server/src/test/java/org/spine3/server/failure/FailureBusShould.java
@@ -26,7 +26,7 @@ import org.spine3.base.CommandContext;
 import org.spine3.base.Commands;
 import org.spine3.base.Failure;
 import org.spine3.base.Failures;
-import org.spine3.base.Subscribe;
+import org.spine3.annotation.Subscribe;
 import org.spine3.change.StringChange;
 import org.spine3.envelope.FailureEnvelope;
 import org.spine3.test.failure.ProjectFailures;
