+++ server/src/test/java/org/spine3/server/reflect/EventSubscriberMethodShould.java
@@ -23,7 +23,7 @@ package org.spine3.server.reflect;
 import com.google.common.testing.NullPointerTester;
 import org.junit.Test;
 import org.spine3.base.EventContext;
-import org.spine3.base.Subscribe;
+import org.spine3.annotation.Subscribe;
 import org.spine3.test.reflect.event.ProjectCreated;
 
 import java.lang.reflect.InvocationTargetException;
